Added a check to make sure we don't try to build the nuget package with incorrect/out of date frameworks inside it.

+++ Binding/FetchFramework.sh
@@ -1,6 +1,20 @@
 #!/bin/sh
 
 carthage update
+
+# If carthage failed.
+if [ $? -ne 0 ]
+then
+   exit 1
+fi
+
 cp -r Carthage/Build/iOS/Mixpanel.framework .
+
+# If we couldn't copy the Mixpanel.framework.
+if [ $? -ne 0 ]
+then
+   exit 1
+fi
+
 rm -rf Carthage
